Use a journal table
We can't commit changes to both etcd and MySQL simultaneously because
etcd may not be available.  We commit intended changes to a
journalling table and feed tehm to etcd in the background.  If any
interruption of connectivity happens we have a journal of updates that
will allow us to bring etcd up to the current stage.